requirements.txt
# Core ML
torch>=2.0.0
pandas>=2.0.0

# Image/Video loading
pillow>=10.0.0
opencv-python>=4.8.0

# Excel support (Cholec80-CVS annotations are in XLSX format)
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
